+++ src/main/java/edu/ie3/datamodel/models/input/thermal/ThermalSinkInput.java
@@ -1,5 +1,5 @@
 /*
- * © 2021. TU Dortmund University,
+ * © 2026. TU Dortmund University,
  * Institute of Energy Systems, Energy Efficiency and Energy Economics,
  * Research group Distribution grid planning and operation
 */
@@ -7,17 +7,18 @@ package edu.ie3.datamodel.models.input.thermal;
 
 import edu.ie3.datamodel.models.OperationTime;
 import edu.ie3.datamodel.models.input.OperatorInput;
+import java.util.Objects;
 import java.util.UUID;
 
-/** Common properties to all thermal sinks */
+/** Common properties to all thermal sinks. */
 public abstract class ThermalSinkInput extends ThermalUnitInput {
   /**
    * @param uuid Unique identifier of a thermal sink input model
    * @param id Identifier of the thermal unit
-   * @param bus Thermal bus, a thermal unit is connected to
+   * @param thermalBus Thermal bus, a thermal unit is connected to
    */
-  ThermalSinkInput(UUID uuid, String id, ThermalBusInput bus) {
-    super(uuid, id, bus);
+  protected ThermalSinkInput(UUID uuid, String id, ThermalBusInput thermalBus) {
+    super(uuid, id, thermalBus);
   }
 
   /**
@@ -25,32 +26,60 @@ public abstract class ThermalSinkInput extends ThermalUnitInput {
    * @param id Identifier of the thermal unit
    * @param operator operator of the asset
    * @param operationTime operation time of the asset
-   * @param bus Thermal bus, a thermal unit is connected to
+   * @param thermalBus Thermal bus, a thermal unit is connected to
    */
-  ThermalSinkInput(
+  protected ThermalSinkInput(
       UUID uuid,
       String id,
       OperatorInput operator,
       OperationTime operationTime,
-      ThermalBusInput bus) {
-    super(uuid, id, operator, operationTime, bus);
+      ThermalBusInput thermalBus) {
+    super(uuid, id, operator, operationTime, thermalBus);
+  }
+
+  @Override
+  public boolean equals(Object o) {
+    if (this == o) return true;
+    if (!(o instanceof ThermalSinkInput that)) return false;
+    return super.equals(o);
+  }
+
+  @Override
+  public int hashCode() {
+    return Objects.hash(super.hashCode());
+  }
+
+  @Override
+  public String toString() {
+    return "ThermalSinkInput{"
+        + "uuid="
+        + getUuid()
+        + ", id="
+        + getId()
+        + ", operator="
+        + getOperator().getUuid()
+        + ", operationTime="
+        + getOperationTime()
+        + ", thermalBus="
+        + getThermalBus().getUuid()
+        + ", additionalInformation="
+        + getAdditionalInformation()
+        + "}";
   }
 
   @Override
   public abstract ThermalSinkInputCopyBuilder<?> copy();
 
-  /**
-   * Abstract class for all builders that build child entities of abstract class {@link
-   * ThermalSinkInput}
-   */
   public abstract static class ThermalSinkInputCopyBuilder<B extends ThermalSinkInputCopyBuilder<B>>
       extends ThermalUnitInputCopyBuilder<B> {
-
     protected ThermalSinkInputCopyBuilder(ThermalSinkInput entity) {
       super(entity);
     }
 
     @Override
     public abstract ThermalSinkInput build();
+
+    @Override
+    protected abstract B thisInstance();
   }
 }